+++ src/common.h
@@ -64,6 +64,7 @@ bool is_type_decl(tree_t t);
 bool is_loop_stmt(tree_t t);
 bool is_design_unit(tree_t t);
 bool is_uninstantiated_package(tree_t pack);
+bool is_uninstantiated_subprogram(tree_t decl);
 tree_t search_decls(tree_t container, ident_t name, int nth);
 bool is_builtin(subprogram_kind_t kind);
 bool is_open_coded_builtin(subprogram_kind_t kind);